[3D] Reading and rendering more game versions

+++ WoWDatabaseEditor.Common/WDE.CMMySqlDatabase/Models/Classic/CreatureClassLevelStat.cs
@@ -9,14 +9,14 @@ namespace WDE.CMMySqlDatabase.Models.Classic
         [Column("Level", IsPrimaryKey = true, PrimaryKeyOrder = 0)] public byte   Level                 { get; set; } // tinyint(4)
         [Column("Class", IsPrimaryKey = true, PrimaryKeyOrder = 1)] public byte   Class                 { get; set; } // tinyint(4)
         [Column("BaseHealthExp0")                                 ] public ushort BaseHp0               { get; set; } // mediumint(8) unsigned
-        [Column("BaseMana")                                       ] public ushort BaseMana              { get; set; } // mediumint(8) unsigned
+        [Column("BaseMana")                                       ] public int BaseMana              { get; set; } // mediumint(8) unsigned
         [Column("BaseDamageExp0")                                 ] public float  DamageBase            { get; set; } // float
         [Column("BaseMeleeAttackPower")                           ] public float  fAttackPower          { get; set; } // float
         [Column("BaseRangedAttackPower")                          ] public float  fRangedAttackPower    { get; set; } // float
-        [Column("BaseArmor")                                      ] public ushort BaseArmor             { get; set; } // mediumint(8) unsigned
+        [Column("BaseArmor")                                      ] public int BaseArmor             { get; set; } // mediumint(8) unsigned
 
-        public ushort AttackPower => (ushort)(fAttackPower > 0.0 ? System.Convert.ToUInt16(fAttackPower) : 0);
-        public ushort RangedAttackPower => (ushort)(fRangedAttackPower > 0.0 ? System.Convert.ToUInt16(fRangedAttackPower) : 0);
+        public int AttackPower => (ushort)(fAttackPower > 0.0 ? System.Convert.ToUInt16(fAttackPower) : 0);
+        public int RangedAttackPower => (ushort)(fRangedAttackPower > 0.0 ? System.Convert.ToUInt16(fRangedAttackPower) : 0);
 
         public uint BaseHp(byte expansion)
         {
